requirements.txt
django
black
pylint
pylint-django
django-dotenv
django-bootstrap4
django-allauth
django-debug-toolbar
# Live reload for development
django-livesync
# Amazon SES (Email)
django-ses
# Stripe
stripe
# Carousel
django-versatileimagefield
# Blurb
django-ckeditor
# Dt Content
django-model-utils
